RIFEX-47 deleting RNS subdomain
* Fixing delete subdomain service and ui.

This affects ui and service layers
